refactor: rename screen files to snake_case
